% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_themes_plots.R
\name{fit_flextable_to_page}
\alias{fit_flextable_to_page}
\title{Scale a flextable to page width}
\usage{
fit_flextable_to_page(ft, pgwidth = 6)
}
\arguments{
\item{ft}{A flextable object.}

\item{pgwidth}{Available table width in inches.}
}
\value{
A flextable object.
}
\description{
Scales table column widths to a specified printable width.
}
\examples{
if (requireNamespace("flextable", quietly = TRUE)) {
  fit_flextable_to_page(flextable::flextable(data.frame(a = 1:3)))
}
}
