% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_clinical_QC.R
\name{print.ekb_clinical_qc}
\alias{print.ekb_clinical_qc}
\title{Print a clinical quality-control summary}
\usage{
\method{print}{ekb_clinical_qc}(x, ...)
}
\arguments{
\item{x}{An ekb_clinical_qc object.}

\item{...}{Additional arguments passed to the underlying function; see Details.}
}
\value{
The input object, invisibly.
}
\description{
Displays the summary of an ekb_clinical_qc object.
}
\details{
Additional arguments are reserved for compatibility with print.
}
\examples{
q <- qc_clinical_data(data.frame(id = 1:3), id = "id", verbose = FALSE)
print(q)
}
