% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_themes_plots.R
\name{theme_ekbmed}
\alias{theme_ekbmed}
\title{Use the ekbMed plot theme}
\usage{
theme_ekbmed(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Base plot font size in points.}

\item{base_family}{Base plot font family.}
}
\value{
A ggplot2 theme.
}
\description{
Provides a compact clinical-reporting ggplot theme.
}
\examples{
theme_ekbmed(base_size = 11)
}
